+++ front/lib/api/assistant/conversation.ts
@@ -1902,7 +1902,6 @@ async function* streamRunAgentEvents(
 
       // All other events that won't impact the database and are related to actions or tokens
       // generation.
-      case "conversation_include_file_params":
       case "dust_app_run_block":
       case "dust_app_run_params":
       case "generation_tokens":
