+++ front/lib/assistant/state/messageReducer.ts
@@ -158,7 +158,6 @@ export function messageReducer(
       return newState;
     }
 
-    case "conversation_include_file_params":
     case "dust_app_run_block":
     case "dust_app_run_params":
     case "search_labels_params":
