+++ front/lib/poke/conversations.ts
@@ -75,7 +75,6 @@ export async function getPokeConversation(
                     };
                     break;
 
-                  case "conversation_include_file_action":
                   case "conversation_list_files_action":
                   case "search_labels_action":
                     // TODO(REASONING TOOL): reasoning_action
