+++ front/types/assistant/conversation.ts
@@ -1,4 +1,3 @@
-import type { ConversationIncludeFileActionType } from "@app/lib/actions/conversation/include_file";
 import type { ConversationListFilesActionType } from "@app/lib/actions/conversation/list_files";
 import type { DustAppRunActionType } from "@app/lib/actions/dust_app_run";
 import type { MCPActionType } from "@app/lib/actions/mcp";
@@ -126,7 +125,6 @@ export type ConfigurableAgentActionType = DustAppRunActionType | MCPActionType;
 
 export type ConversationAgentActionType =
   | ConversationListFilesActionType
-  | ConversationIncludeFileActionType;
 
 export type AgentActionType =
   | ConfigurableAgentActionType
@@ -140,7 +138,6 @@ export type AgentMessageStatus =
   | "cancelled";
 
 export const ACTION_RUNNING_LABELS: Record<AgentActionType["type"], string> = {
-  conversation_include_file_action: "Reading file",
   conversation_list_files_action: "Listing files",
   dust_app_run_action: "Running App",
   search_labels_action: "Searching labels",
